requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
scikit-learn
joblib
sentence-transformers
torch
transformers
tqdm
openai-whisper
yt-dlp
opencv-python
Pillow
ffmpeg-python